Add script to query GH project popularity/activity.
This commit updates README.md to include the number of stars a project
has, and days since last activity. These are represented using the
unicode symbols ★ and ⧗.

For example, "Spire ★ 737 ⧗ 2" means that Spire has 737 "stargazers"
and has had activity within the last two days. In other words, the
first number corresponds with popularity (larger is better) and the
second number corresponds with activity (lower is better). We bold the
titles of projects that have 500 or more stars.

We look for a GitHub Personal Access Token stored in ./access-token.
If found, this will allow more queries to be made (the limit for
anonymous users is 60/hour).

There are two ways to run the script:

  ./metadata.py

    This will update projects lacking metadata entirely, leaving
    everything else as-is. In this mode, no Personal Access Token
    is needed if only a few new projects are being added.

  ./metadata.py --update

    This will update all projects. We should run this periodically
    since projects will be adding stars over time, and projects'
    activity levels will change over time as well. This mode requires
    a Personal Access Token, since we already have 120 repositories
    (double the hourly limit for anonymous API access).

The script also supports a --fake option to allow testing without
accessing GitHub's API.

README.md has not been updated yet; the author seems to have exceeded
GitHub's authenticated rate limit during testing.